Added smartSkip + others
smartSkip will skip the song and move to the user to the lockskip
position. Lockskips will be triggered if the waitlist is full. You can
toggle smartSkip in the bot settings.

Voteskip has also been improved with smartSkip, it will give the user a
second chance if the majority of the song hasn't yet been played by
moving the user to the lockskip position specified in the bot settings.

Autowoot can now be toggled in the bot settings.

Improved Autoskip.